pyproject.toml
[project]
name = "object-detection-mobile"
version = "1.0.0"
description = "Mobile object detection system with TensorFlow Lite and FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "tensorflow>=2.15.0",
    "tensorflow-hub>=0.15.0",
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    # Data processing
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Experiment tracking
    "mlflow>=2.9.0",
    # Database (optional)
    "psycopg2-binary>=2.9.9",
    # HTTP client
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "streamlit>=1.52.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
